First portion of code and comment restructuring
