+++ components/src/jsMain/kotlin/dev/fritz2/components/alert.kt
@@ -1,180 +1,26 @@
 package dev.fritz2.components
 
+import dev.fritz2.components.alert.AlertComponent
 import dev.fritz2.components.validation.ComponentValidationMessage
 import dev.fritz2.components.validation.Severity
 import dev.fritz2.dom.html.RenderContext
 import dev.fritz2.styling.StyleClass
-import dev.fritz2.styling.div
 import dev.fritz2.styling.params.BasicParams
-import dev.fritz2.styling.params.BoxParams
-import dev.fritz2.styling.params.Style
-import dev.fritz2.styling.span
-import dev.fritz2.styling.style
-import dev.fritz2.styling.theme.*
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOf
+
 
 /**
- * A component to display an alert consisting of an icon, title and description.
- * Different styles based on severities are supported, as well as a number of different layout options.
+ * Creates an alert and renders it right away.
  *
- * Currently the following severities are available:
- * - Info
- * - Success
- * - Warning
- * - Error
- * Specifying a severity will change the alert's color scheme based on the colors defined in the application theme as
- * well as the icon displayed. If no severity is specified, 'info' will be used by default.
- * The alert's icon can manually be overridden by setting the respective dsl property.
- *
- * Additionally, a number of different layout options are available. These are:
- * - 'subtle': A subtle style using different shades of the severity's base color defined in the application theme.
- * - 'solid': A solid style using the severity's color from the application theme and a solid white color for the icon,
- * text and decorations.
- * - 'leftAccent': A variation of the 'subtle' variant with a decoration element at the top.
- * - 'topAccent': A variation of the 'subtle' variant with a decoration element on the left.
- * - 'ghost': This variant does not have any decoration besides the icon (no background, similar to 'ghost' in push-buttons)
- * If no variant is specified, 'solid' is used by default.
- *
- * Usage examples:
+ * Usage example:
  * ```
  * alert {
  *     title("Alert")
  *     content("This is an alert.")
  *     severity { info }
  * }
- *
- * alert {
- *     title("Alert")
- *     content("This is an alert.")
- *     severity { error }
- *     variant { leftAccent }
- * }
- *
- * alert {
- *     title("Alert")
- *     content("This is an alert.")
- *     icon { fritz2 }
- * }
  * ```
- */
-open class AlertComponent : Component<Unit> {
-
-    companion object {
-        private val alertCss = style("alert") {
-            display { flex }
-            alignItems { center }
-        }
-
-        private val alertContentCss = style("alert-content") {
-            display { inlineBlock }
-            width { "100%" }
-            lineHeight { "1.2em" }
-        }
-    }
-
-    enum class Variant {
-        SOLID, SUBTLE, LEFT_ACCENT, TOP_ACCENT, GHOST
-    }
-
-    object VariantContext {
-        val solid = Variant.SOLID
-        val subtle = Variant.SUBTLE
-        val leftAccent = Variant.LEFT_ACCENT
-        val topAccent = Variant.TOP_ACCENT
-        val ghost = Variant.GHOST
-    }
-
-    val icon = ComponentProperty<(Icons.() -> IconDefinition)?>(null)
-    val severity = ComponentProperty<(AlertSeverities.() -> AlertSeverity)> { info }
-    val variant = ComponentProperty<VariantContext.() -> Variant> { solid }
-    val size = ComponentProperty<FormSizes.() -> Style<BasicParams>> { normal }
-    val stacking = ComponentProperty<AlertStacking.() -> Style<BasicParams>> { separated }
-
-    private val actualIcon: IconDefinition
-        get() = icon.value?.invoke(Theme().icons)
-            ?: severity.value(Theme().alert.severities).icon
-
-
-    private var title: (RenderContext.() -> Unit)? = null
-
-    fun title(value: RenderContext.() -> Unit) {
-        title = value
-    }
-
-    fun title(value: Flow<String>) {
-        title {
-            span({
-                margins { right { smaller } }
-                fontWeight { bold }
-            }) {
-                value.asText()
-            }
-        }
-    }
-
-    fun title(value: String) = title(flowOf(value))
-
-
-    private var content: (RenderContext.() -> Unit)? = null
-
-    fun content(value: RenderContext.() -> Unit) {
-        content = value
-    }
-
-    fun content(value: Flow<String>) {
-        content {
-            span { value.asText() }
-        }
-    }
-
-    fun content(value: String) = content(flowOf(value))
-
-
-    override fun render(
-        context: RenderContext,
-        styling: BoxParams.() -> Unit,
-        baseClass: StyleClass,
-        id: String?,
-        prefix: String,
-    ) {
-        context.apply {
-            div({
-                this@AlertComponent.size.value(Theme().alert.sizes)()
-                this@AlertComponent.stacking.value(Theme().alert.stacking)()
-                when (this@AlertComponent.variant.value(VariantContext)) {
-                    Variant.SOLID -> Theme().alert.variants.solid
-                    Variant.SUBTLE -> Theme().alert.variants.subtle
-                    Variant.LEFT_ACCENT -> Theme().alert.variants.leftAccent
-                    Variant.TOP_ACCENT -> Theme().alert.variants.topAccent
-                    Variant.GHOST -> Theme().alert.variants.ghost
-                }.invoke(this, this@AlertComponent.severity.value(Theme().alert.severities).colorScheme)
-                styling()
-            }, alertCss) {
-                div({
-                    display { inlineFlex }
-                    css("margin-right: var(--al-icon-margin)")
-                }) {
-                    icon({
-                        css("width: var(--al-icon-size)")
-                        css("height: var(--al-icon-size)")
-                    }) {
-                        fromTheme { this@AlertComponent.actualIcon }
-                    }
-                }
-
-                div(alertContentCss.name) {
-                    this@AlertComponent.title?.invoke(this)
-                    this@AlertComponent.content?.invoke(this)
-                }
-            }
-        }
-    }
-}
-
-
-/**
- * Creates an alert and renders it right away.
+ *
+ * @see AlertComponent
  *
  * @param styling a lambda expression for declaring the styling of the toast using fritz2's styling DSL
  * @param baseClass optional CSS class that should be applied to the toast element
@@ -194,10 +40,21 @@ fun RenderContext.alert(
 
 
 /**
- * Convenience extension to display a [ComponentValidationMessage] as an alert.
+ * Convenience extension function to display a [ComponentValidationMessage] as an alert.
+ *
  * The alert's severity and content are determined from the validation message's properties.
  * Custom styling via the [styling] parameter is optionally supported, as well as any other customization via the
  * [build]-lambda.
+ *
+ * Usage example:
+ * ```
+ * val message = infoMessage("SomeId", "This is a simple info message.")
+ *
+ * // within some `RenderContext` so `this` refers to it
+ * message.asAlert(this)
+ * ```
+ *
+ * @see AlertComponent
  *
  * @param renderContext RenderContext to render the alert in
  * @param styling a lambda expression for declaring the styling of the toast using fritz2's styling DSL
@@ -225,9 +82,22 @@ fun ComponentValidationMessage.asAlert(
 
 /**
  * Convenience extension to display a [ComponentValidationMessage] as an alert.
+ *
  * The alert's severity and content are determined from the validation message's properties.
- * Customization of the underlying [AlertComponent] is supported via the [build]-lambda. Custom styling can be applyied
- * via the overloaded [ComponentValidationMessage.asAlert] method.
+ * Customization of the underlying [AlertComponent] is supported via the [build]-lambda.
+ *
+ * This variant does not offer custom styling. The latter can be applied via the overloaded
+ * [ComponentValidationMessage.asAlert] method.
+ *
+ * Usage example:
+ * ```
+ * val message = infoMessage("SomeId", "This is a simple info message.")
+ *
+ * // within some `RenderContext` so `this` refers to it
+ * message.asAlert(this)
+ * ```
+ *
+ * @see AlertComponent
  *
  * @param renderContext RenderContext to render the alert in
  * @param build a lambda expression for setting up the component itself
